% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_13datasets.R
\docType{data}
\name{all_13datasets}
\alias{all_13datasets}
\title{Example Dataset for testing RaMBat
Dataset contains an expression matrix where rows are genes and columns are samples; this dataset is the combination of GSE73038, GSE10327, GSE21140, GSE67850, GSE50765, GSE49243, GSE30074, GSE37382, GSE74195, GSE62803, GSE41842, GSE50161, GSE12992
Dataset required by RaMBat}
\format{
An object of class \code{data.frame} with 15417 rows and 508 columns.
}
\usage{
all_13datasets
}
\description{
Example Dataset for testing RaMBat
Dataset contains an expression matrix where rows are genes and columns are samples; this dataset is the combination of GSE73038, GSE10327, GSE21140, GSE67850, GSE50765, GSE49243, GSE30074, GSE37382, GSE74195, GSE62803, GSE41842, GSE50161, GSE12992
Dataset required by RaMBat
}
\keyword{datasets}
