% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_rank_t_genes.R
\docType{data}
\name{all_rank_t_genes}
\alias{all_rank_t_genes}
\title{Differentially ranked genes selected form example dataset GSE85217
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat}
\format{
An object of class \code{list} of length 4.
}
\usage{
all_rank_t_genes
}
\description{
Differentially ranked genes selected form example dataset GSE85217
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat
}
\keyword{datasets}
