% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_reversed_gp_genes.R
\docType{data}
\name{all_reversed_gp_genes}
\alias{all_reversed_gp_genes}
\title{extracted reversed gene ratios from example dataset (GSE85217) for each subtype based on genes selected from RCA() function
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat}
\format{
An object of class \code{list} of length 4.
}
\usage{
all_reversed_gp_genes
}
\description{
extracted reversed gene ratios from example dataset (GSE85217) for each subtype based on genes selected from RCA() function
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat
}
\keyword{datasets}
