% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneRankAnalysis.R
\name{GRA}
\alias{GRA}
\title{extract differetially ranked genes for each subtype}
\usage{
GRA(data, sampAnnot)
}
\arguments{
\item{data}{the original dataset we extract features}

\item{sampAnnot}{sampAnnot is the annotation file for each samples in data}
}
\value{
a global object all_rank_t_genes
}
\description{
this function is used to extratc differetially ranked genes from dataset for each subtype
}
\examples{
all_rank_t_genes<- RCA(GSE85217, sampAnnote_GSE85217)
}
