% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE85217.R
\docType{data}
\name{GSE85217}
\alias{GSE85217}
\title{Example Dataset for training RaMBat
Dataset contains an expression matrix where rows are genes and columns are samples; this dataset is GSE85217
Dataset required by RaMBat}
\format{
An object of class \code{data.frame} with 21039 rows and 763 columns.
}
\usage{
GSE85217
}
\description{
Example Dataset for training RaMBat
Dataset contains an expression matrix where rows are genes and columns are samples; this dataset is GSE85217
Dataset required by RaMBat
}
\keyword{datasets}
