% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_RANK_GP.R
\docType{data}
\name{MB_RANK_GP}
\alias{MB_RANK_GP}
\title{extracted important gene ratios from example dataset (GSE85217) for each subtype based on genes ratios selected from RRA() function
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat}
\format{
An object of class \code{list} of length 4.
}
\usage{
MB_RANK_GP
}
\description{
extracted important gene ratios from example dataset (GSE85217) for each subtype based on genes ratios selected from RRA() function
Dataset contains four list for SHH, WNT,Group3, Group4, respectively
Dataset required by RaMBat
}
\keyword{datasets}
