% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSubtyping_predict.R
\name{predMB}
\alias{predMB}
\title{predict medulloblastoma (MB) subtype}
\usage{
predMB(data)
}
\arguments{
\item{data}{the given dataset we need to predit subtype information based on extracted features}
}
\value{
a global object myMat
}
\description{
this function is used to subtype given samples. Based on gene ratios we selected from previous steps, for each sample, we predict its subtype
}
\examples{
myMat<-preMB(data)
}
