% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversalRatioAnalysis.R
\name{RRA}
\alias{RRA}
\title{extract reversed gene ratio}
\usage{
RRA(data, sampAnnote, all_rank_t_genes)
}
\arguments{
\item{data}{the original dataset we extract features}

\item{sampAnnote}{sampAnnot is the annotation file for each samples in data}

\item{all_rank_t_genes}{all_rank_t_genes is the differentially ranked genes selected in previous RCA() function}
}
\value{
a global object all_reversed_gp_genes
}
\description{
this function is used to extract reversed gene ratios from dataset for each subtype based on genes selected from RCA() function
}
\examples{
all_reversed_gp_genes<-RRA(GSE85217, sampAnnote_GSE85217, all_rank_t_genes)
}
