% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_13.R
\docType{data}
\name{samp_13}
\alias{samp_13}
\title{Annotation file for example dataset all_13datasets  for testing RaMBat
Dataset contains three columns, including 'gsm, Subtype, dataset'
Dataset required by RaMBat}
\format{
An object of class \code{data.frame} with 508 rows and 3 columns.
}
\usage{
samp_13
}
\description{
Annotation file for example dataset all_13datasets  for testing RaMBat
Dataset contains three columns, including 'gsm, Subtype, dataset'
Dataset required by RaMBat
}
\keyword{datasets}
