% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampAnnot_GSE85217.R
\docType{data}
\name{sampAnnot_GSE85217}
\alias{sampAnnot_GSE85217}
\title{Annotation file for example dataset GSE85217  for training RaMBat
Dataset contains three columns, including 'Sample, Subtype'
Dataset required by RaMBat}
\format{
An object of class \code{data.frame} with 763 rows and 2 columns.
}
\usage{
sampAnnot_GSE85217
}
\description{
Annotation file for example dataset GSE85217  for training RaMBat
Dataset contains three columns, including 'Sample, Subtype'
Dataset required by RaMBat
}
\keyword{datasets}
