% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSubtyping_test.R
\name{testMB}
\alias{testMB}
\title{predict medulloblastoma (MB) subtype}
\usage{
testMB(data)
}
\arguments{
\item{data}{the example data (all_13datasets) to test the performance of RaMBat}
}
\value{
a global object pred_result
}
\description{
this function is used to subtype given samples based on gene ratios we selected from previous steps, for each sample, we predict its subtype
}
\examples{
pred_result<-testMB(all_datasets, MB_RANK_GP, sampAnnote_all)
}
