% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneRankAnalysis.R
\name{GRA}
\alias{GRA}
\title{extract differentially ranked genes for each subtype}
\usage{
GRA(data, sampAnnot)
}
\arguments{
\item{data}{the training dataset we extract features}

\item{sampAnnot}{sampAnnot is the annotation file for each samples in training dataset}
}
\value{
a global object 'all_rank_t_genes'
}
\description{
this function is used to extract differentially ranked genes from dataset for each subtype
}
\examples{
all_rank_t_genes<- GRA(GSE85217, sampAnnote_GSE85217)
}
