% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversalRatioAnalysis.R
\name{RRA}
\alias{RRA}
\title{extract reversed gene ratio}
\usage{
RRA(data, sampAnnot, all_rank_t_genes)
}
\arguments{
\item{data}{the training dataset we extract gene ratios (GERs)}

\item{sampAnnot}{sampAnnot is the annotation file for each samples in training dataset}

\item{all_rank_t_genes}{all_rank_t_genes is the differentially ranked genes selected in previous GRA() function}
}
\value{
a global object 'all_reversed_gp_genes'
}
\description{
this function is used to extract reversed gene ratios (GERs) from dataset for each subtype based on genes selected from GRA() function
}
\examples{
all_reversed_gp_genes<-RRA(GSE85217, sampAnnote_GSE85217, all_rank_t_genes)
}
