% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSubtyping_test.R
\name{testMB}
\alias{testMB}
\title{predict medulloblastoma (MB) subtype}
\usage{
testMB(data, MB_RANK_GP, sampAnnot)
}
\arguments{
\item{data}{the example independent test dataset (all_13datasets) to evaluate the performance of RaMBat}

\item{MB_RANK_GP}{the unique gene ratios (GERs) selected from the training dataset}

\item{sampAnnot}{the sampAnnot is the annotation file for each sample in training dataset}
}
\value{
a global object 'Pred_performance'
}
\description{
this function is used to subtype given samples based on gene ratios we selected from previous steps, for each sample, we predict its subtype
}
\examples{
Pred_performance<-testMB(all_13datasets, MB_RANK_GP, samp_13)
}
